Post merge fixes.

+++ src/main/java/com/teamsimplyrs/prismaarcanum/api/spell/spells/ventus/WindPool.java
@@ -26,7 +26,7 @@ public class WindPool extends AbstractSpell {
 
     public static final int tier = 1;
     public static final int basicManaCost = 0;
-    public static final float basicCooldown = 1f;
+    public static final int basicCooldown = 1;
 
     public static final boolean hasEvolution = true;
